refactor timer impl to allocate per-Future state
This enables us to make a lot more scheduling code trivially sound as the complexity of the SQ and CQ grows under cancel-and-drop semantics coupled with manual polling and eager-dropping Futures. Users can create streams of CQEs out from under us which must be ignored by the scheduler.

Cancellation is also not guaranteed to complete inline with ring submission when the work is dispatched to the io-wq, creating a schism in how these I/O objects are handled.

Using an allocation per Future enables us to much more easily guarantee we handle complex  scheduling and user interaction with the library.