add basic buffer replenishment scheme
We fixed several things here.

We need to use a CQ advance guard in the main event loop to handle the case where the library internally panics. If we don't mark the CQEs as seen they'll be in the CQ when the RunGuard starts cleaning up which leads to immediate UAFs.

We also correct a memory leak where `release_op` was never called for a recv op cancelled by the I/O object destructor.

More importantly, we add a test that ensures we process the recv bundles correctly and preserve ordering, even in the face of buffer replenishment schemes.