add client version of tcp timeout testing
We also have to fudge the numbers in the sleep a bit as well. The tests can be kind of flaky in debug builds and we do wind up actually running into the problematic cases where actual timeout for a given t is somewhere between `[t, 2 * t)`. So, we adjust the sleep to be `t` regardless and this seems to behave properly.