replace local task queue with LocalWaker
Libraries like futures outright require the usage of a Waker to function correctly, i.e. FuturesUnordered won't work without it. This is because these libraries will create wrapper Wakers that mutate state they own when they're polled. It's up to us to store the provided LocalWaker from a context as that's what futures polls our Futures with.

The SpawnFuture can probably be optimized to use the LocalWaker as well.