add BorrowedBufs type
Update our multishot recv operation to return borrows to the backing
buffer sequence used when setting up the buf_ring. This means that
buffers are returned back to the ring during Drop, which can make
userland code potentially susceptible to starvation.

By returning borrows to the buffers, we also avoid excessive
allocations which aren't problematic now in benchmarks but can
potentially wreak havoc for long-standing programs as the heap begins
to fragment.