update Tasks to store results inline
Currently, SpawnFutures and SpawnValues share ownership of a
heap-allocated value that the top-level Future can write to when it's
done. To avoid allocating to the heap, the top-level Task now reserves
space for the SpawnValue and the WrapperFuture writes to that location
when the wrapped top-level Future finally completes.

The written to SpawnValue is considered as a part of the strong portion
of the task. This is to make sure that the Future's return value has
no Send requirements.