update strong count in Task to be non-atomic
Because Task isn't Send, there's no reason for the strong count to be
an AtomicU64. We micro-optimize by removing the usage of the atomic
which also dramatically simplifies the upgrade loop as well. We also
update the TaskHeader to be properly Send because we destroy it when
the weak count reaches zero. This doesn't make the code more sound but
helps make the code more self-documenting.