update Timer to no longer allocate operation state
In the name of removing all extraneous allocations from the codebase,
the Timer class is now updated to no longer hold its operation state
via a Box. Because of this, we can no longer leak the allocation and
then immediately reuse the I/O object. The code now forces the user to
process the corresponding CQE for the I/O object before the same I/O
operation can begin again.

This change only affects eager-dropped Futures which really only
happens when using the select! macro. Hopefully the yield requirement
does not prove prohibitive in practice. Futures that are forgotten mean
that the I/O object can never be used for that I/O again as the CQE
processing loop will not mark the operation done.