update test to give grace period for straggling CQES
We only release resources once we've seen a corresponding CQE. In some
cases, it's impossible to guarantee that we're fast or slow enough such
that all CQEs are handled. This test in particular is problematic
because it attempts to terminate several multishot timeouts at once.
Giving the test a grace period helps ensure that the CQEs populate
before we exit.