update recv buffer sequences to no longer allocate
In continued efforts to remove as many allocations as possible, we
update the BufGroup structure to store a vector of headers which we can
use to generate a doubly-linked list of buffers in-place without the
need to ever reallocate during runtime.

This also solves our space issues with the current borrowed buffer
implementation which created a rather large structure that was passed
on the stack.