fix sqe stability soundness hole
io_uring requires that data provided to SQEs be stable until submission
has finished. This means that if we lower the size of our SlotMap to
something realistic for user machines and it actually starts
reallocating when we insert new operations into it, the data for the
SQEs is no longer stable so operations will complete with an EFAULT.

This can be readily fixed by simply submitting the ring when the
SlotMap is full. We update the code to require going through an
interface that ensures this.