fix bug in fd reclamation
Returning a file descriptor back to the ring must only happen once the
associated CQE has been seen. If we schedule the SQE for closing the
descriptor, it leaves us vulnerable to spurious close() calls because
the SQ isn't slated for immediate work and the runtime is free to hand
out the fd erroneously.

This currently would only manifest for the `tcp::Acceptor` class which
manually maps in a normal fd into the ring, but would become much more
apparent if the runtime supported mapping in all types of fds.