avoid double-close when possible
Close-on-drop is a mostly good design but unconditional close-on-drop
is strictly inefficient and gratuitous. We update the StreamImpl to
check if the socket has been previously closed and if it has, we skip
scheduling the DropClose and return the fd directly back to the runtime.

Because CQEs lack visibility into the StreamImpl structure, we can only
mark a socket as having been closed from the CloseFuture which means
there are edge-cases we can't avoid where we will schedule an
unnecessary DropClose and reap a CQE with a EBADF.