amortize eventfd write during wakeup
We can dramatically reduce the amount of writes to the eventfd during
wakeup if we use an atomic flag before writing. We know that we won't
miss any task scheduling if we check the queues before blocking on
io_uring_submit_and_wait. This is because any wakeup scheduled after
the atomic flag has been set is guaranteed to generate a wakeup.