fix scheduling bug
SpawnFutures now participate in the lifetime of their backing task. This
means that it's possible for a SpawnFuture to keep a Task's strong
portion alive even though the task itself has completed. This means that
Wakers associated with the task will wind up scheduling a Task with an
active strong count which the runtime assumes is a valid Future, which
leads to polling completed Futures and resulting in a panic.

We update the code to store a boolean denoting whether or not the task
has completed and eschew scheduling if that's the case.