patch soundness hole in TaskHeader
The TaskHeader was missing a Sync requirement which meant that with the
recent refactors enabling the usage of Cell for dangling pointers to
other tasks, code from other threads could introduce data races without
the usage of `unsafe` which is a soundness hole.

To remedy this, we add a test that requires Sync for the TaskHeader and
then update the code to use SyncUnsafeCell internally so that all
accesses to the non-atomic data are now `unsafe`.