add SO_REUSEPORT support
We introduce a new AcceptorOpts struct which can be used for configuring
the socket options before we call bind() on the listening socket in the
Acceptor implementation.

We also dramatically clean up the implementation to dispatch through a
single function which reduces code duplication.

The provided tests wind up creating 16 client connections when only 2
are required because we can't make a hard guarantee that the kernel will
load balance across both listening sockets but this ratio seems to work
reliably enough in practice when running the test suite multiple times.