add send_subspan
We extend the TCP stream class to support incremental sending of the
backing message. Panics on an empty range because it's almost certainly
always the result of a bug. Taking the `std::ops::Range` argument first
enables us to use `(n..buf.len())` in the function call, otherwise a
borrowing error is generated under the current implementation.